Compartimentize into the Authentication component
